scroll to top bug fix
